requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.weakref==1.0.post1
backports.zoneinfo==0.2.1
bleach==5.0.0
certifi==2022.5.18.1
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.1
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.6
django-cors-headers==3.12.0
django-dotenv==1.4.2
django-filter==21.1
django-nine==0.2.5
django-phonenumber-field==6.3.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
model-bakery==1.5.0
numpy==1.23.1
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
phonenumbers==8.12.53
Pillow==9.1.1
pkg_resources==0.0.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-memcached==1.59
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.11
webencodings==0.5.1
